Data creation
First, understand the code based on the design of the database, and generate the ER diagram of the SQLAlchemy model (Mermaid Live Editor). The graphic can be seen directly. Then, create a CSV file by matching each field based on the graph.Specify the expression of fields and continuously modify and improve them to make the data logical.